tests/wpt/web-platform-tests/shadow-dom/slots-imperative-api-slotchange.tentative.html

<!DOCTYPE html>
<title>Shadow DOM: Imperative Slot API slotchange event</title>
<meta name="author" title="Yu Han" href="mailto:yuzhehan@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://dom.spec.whatwg.org/#signaling-slot-change">
<script src="resources/shadow-dom.js"></script>

<div id="test_slotchange">
  <div id="host">
    <template id="shadow_root" data-mode="open" data-slot-assignment="manual">
      <slot id="s1"></slot>
      <slot id="s2"></slot>
    </template>
    <div id="c1"></div>
    <div id="c2"></div>
  </div>
  <div id="c4"></div>
</div>

<script>
function getDataCollection() {
  return {
    s1EventCount: 0,
    s2EventCount: 0,
    s1ResolveFn: null,
    s2ResolveFn: null
  }
}

function setupShadowDOM(test, data) {
  let tTree = createTestTree(test_slotchange);
  tTree.s1.addEventListener('slotchange', (event) => {
    if (!event.isFakeEvent) {
      test.step(function () {
          assert_equals(event.type, 'slotchange', 'slotchange event\'s type must be "slotchange"');
          assert_equals(event.target, tTree.s1, 'slotchange event\'s target must be the slot element');
          assert_equals(event.relatedTarget, undefined, 'slotchange must not set relatedTarget');
      });
      data.s1EventCount++;
    }
    data.s1ResolveFn();
  });
  tTree.s2.addEventListener('slotchange', (event) => {
    if (!event.isFakeEvent) {
      test.step(function () {
          assert_equals(event.type, 'slotchange', 'slotchange event\'s type must be "slotchange"');
          assert_equals(event.target, tTree.s2, 'slotchange event\'s target must be the slot element');
          assert_equals(event.relatedTarget, undefined, 'slotchange must not set relatedTarget');
      });
      data.s2EventCount++;
    }
    data.s2ResolveFn();
  });
  return tTree;
}

function monitorSlots(data) {
    const s1Promise = new Promise((resolve, reject) => {
      data.s1ResolveFn = resolve;
    });
    const s2Promise = new Promise((resolve, reject) => {
      data.s2ResolveFn = resolve;
    });
    return [s1Promise, s2Promise];
}
</script>

<script>
// Tests:
async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise, s2Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1]);
  tTree.s2.assign([tTree.c2]);

  assert_equals(data.s1EventCount, 0, 'slotchange event must not be fired synchronously');
  assert_equals(data.s2EventCount, 0);

  Promise.all([s1Promise, s2Promise]).then(test.step_func_done(() => {
    assert_equals(data.s1EventCount, 1);
    assert_equals(data.s2EventCount, 1);
  }));
}, 'slotchange event must not fire synchronously.');

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise, s2Promise] = monitorSlots(data);

  assert_throws_dom('NotAllowedError', () => { tTree.s1.assign([tTree.c4]); });
  tTree.s2.assign([]);
  tTree.host.insertBefore(tTree.c4, tTree.c1);

  Promise.all([s1Promise, s2Promise]).then(test.step_func_done(() => {
    assert_equals(data.s1EventCount, 0);
    assert_equals(data.s2EventCount, 0);
  }));

  // use fake event to trigger event handler.
  let fakeEvent = new Event('slotchange');
  fakeEvent.isFakeEvent = true;
  tTree.s1.dispatchEvent(fakeEvent);
  tTree.s2.dispatchEvent(fakeEvent);
}, 'slotchange event should not fire when assignments do not change assignedNodes.');

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1, tTree.c2]);

  s1Promise.then(test.step_func(() => {
    assert_equals(data.s1EventCount, 1);

    [s1Promise] = monitorSlots(data);
    tTree.s1.assign([tTree.c1, tTree.c2]);
    tTree.s1.assign([tTree.c2, tTree.c1, tTree.c1, tTree.c2, tTree.c2]);

    s1Promise.then(test.step_func_done(() => {
      assert_equals(data.s1EventCount, 1);
    }));

    let fakeEvent = new Event('slotchange');
    fakeEvent.isFakeEvent = true;
    tTree.s1.dispatchEvent(fakeEvent);
  }));

}, 'slotchange event should not fire when same node is assigned.');

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise, s2Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1]);
  tTree.s2.assign([tTree.c2]);

  Promise.all([s1Promise, s2Promise]).then(test.step_func_done(() => {
    assert_equals(data.s1EventCount, 1);
    assert_equals(data.s2EventCount, 1);
  }));
}, "Fire slotchange event when slot's assigned nodes changes.");

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise, s2Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1]);

  s1Promise.then(test.step_func(() => {
    assert_equals(data.s1EventCount, 1);

    [s1Promise, s2Promise] = monitorSlots(data);
    tTree.s2.assign([tTree.c1]);

    Promise.all([s1Promise, s2Promise]).then(test.step_func_done(() => {
      assert_equals(data.s1EventCount, 2);
      assert_equals(data.s2EventCount, 1);
    }));
  }));
}, "Fire slotchange event on previous slot and new slot when node is reassigned.");

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1]);

  s1Promise.then(test.step_func(() => {
    assert_equals(data.s1EventCount, 1);

    [s1Promise] = monitorSlots(data);
    tTree.s1.assign([]);

    s1Promise.then(test.step_func_done(() => {
      assert_equals(data.s1EventCount, 2);
    }));
  }));
}, "Fire slotchange event on node assignment and when assigned node is removed.");

async_test((test) => {
  const data = getDataCollection();
  let tTree = setupShadowDOM(test, data);
  let [s1Promise] = monitorSlots(data);

  tTree.s1.assign([tTree.c1, tTree.c2]);

  s1Promise.then(test.step_func(() => {
    assert_equals(data.s1EventCount, 1);

    [s1Promise] = monitorSlots(data);
    tTree.s1.assign([tTree.c2, tTree.c1]);

    s1Promise.then(test.step_func_done(() => {
      assert_equals(data.s1EventCount, 2);
    }));
  }));
}, "Fire slotchange event when order of assigned nodes changes.");
</script>
